+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+## Added
+- Link to JetBrains IDE plugin (WebStorm, IDEA, ...)
+
+## Removed
+- Deprecated bootstrap options
+
 ## [2.0.0-4]
 ## Fixed
 - `getInstanceByToken` returns wrapped instance
@@ -57,6 +64,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 1.0.0
 - Initial release
 
+[Unreleased]: https://github.com/L2jLiga/fastify-decorators/compare/v2.0.0-4...HEAD
 [2.0.0-4]: https://github.com/L2jLiga/fastify-decorators/compare/v2.0.0-3...v2.0.0-4
 [2.0.0-3]: https://github.com/L2jLiga/fastify-decorators/compare/v2.0.0-2...v2.0.0-3
 [2.0.0-2]: https://github.com/L2jLiga/fastify-decorators/compare/v2.0.0-1...v2.0.0-2